Separated Validation and Builder Syntax.
Moved related validation methods into their own class.
Created a proper Builder to chain validation.
Setup a way to log which tests failed. May want to revisit to provide a
cleaner solution.
Removed a lot of unnecessary conditional statements.
Converted the project to Maven to allow it to be used as a dependency.
Created unit tests.